Move IRefactorNotifyService down to the features layer
