+++ src/EditorFeatures/Core/Shared/Extensions/IRefactorNotifyServiceExtensions.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections.Generic;
+using Microsoft.CodeAnalysis.Rename;
 
 namespace Microsoft.CodeAnalysis.Editor.Shared.Extensions;
