+++ src/EditorFeatures/TestUtilities/RenameTracking/MockRefactorNotifyService.cs
@@ -6,6 +6,7 @@
 
 using System.Collections.Generic;
 using System.Runtime.InteropServices;
+using Microsoft.CodeAnalysis.Rename;
 
 namespace Microsoft.CodeAnalysis.Editor.UnitTests.RenameTracking;
