+++ src/Tools/ExternalAccess/Razor/Features/IRazorRefactorNotifyService.cs
@@ -3,7 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections.Generic;
-using Microsoft.CodeAnalysis.Editor;
+using Microsoft.CodeAnalysis.Rename;
 
 namespace Microsoft.CodeAnalysis.ExternalAccess.Razor;
