+++ src/Tools/ExternalAccess/Razor/Features/RazorRefactorNotifyWrapper.cs
@@ -5,8 +5,8 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
-using Microsoft.CodeAnalysis.Editor;
 using Microsoft.CodeAnalysis.Host.Mef;
+using Microsoft.CodeAnalysis.Rename;
 
 namespace Microsoft.CodeAnalysis.ExternalAccess.Razor;
