+++ src/VisualStudio/CSharp/Impl/CodeModel/CSharpCodeModelServiceFactory.cs
@@ -6,11 +6,11 @@ using System;
 using System.Collections.Generic;
 using System.Composition;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Editor;
 using Microsoft.CodeAnalysis.Editor.Shared.Utilities;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.CodeAnalysis.Options;
+using Microsoft.CodeAnalysis.Rename;
 using Microsoft.VisualStudio.LanguageServices.Implementation.CodeModel;
 
 namespace Microsoft.VisualStudio.LanguageServices.CSharp.CodeModel;
