+++ src/VisualStudio/Core/Test/Venus/CSharpContainedLanguageSupportTests.vb
@@ -5,8 +5,8 @@
 Imports System.Threading
 Imports Microsoft.CodeAnalysis
 Imports Microsoft.CodeAnalysis.Collections
-Imports Microsoft.CodeAnalysis.Editor
 Imports Microsoft.CodeAnalysis.Editor.UnitTests.Extensions
+Imports Microsoft.CodeAnalysis.Rename
 Imports Microsoft.CodeAnalysis.Test.Utilities
 Imports Microsoft.VisualStudio.LanguageServices.CSharp.Utilities
 Imports Microsoft.VisualStudio.LanguageServices.Implementation.Venus
